Extract video repository
