+++ src/main/tv/codely/scala_http_api/module/video/domain/VideoRepository.scala
@@ -1,0 +1,7 @@
+package tv.codely.scala_http_api.module.video.domain
+
+trait VideoRepository {
+  def all(): Seq[Video]
+
+  def save(video: Video): Unit
+}
