+++ src/main/tv/codely/scala_http_api/module/video/infrastructure/dependency_injection/VideoModuleDependencyContainer.scala
@@ -1,7 +1,10 @@
 package tv.codely.scala_http_api.module.video.infrastructure.dependency_injection
 
 import tv.codely.scala_http_api.module.video.application.VideosSearcher
+import tv.codely.scala_http_api.module.video.infrastructure.repository.InMemoryVideoRepository
 
 final class VideoModuleDependencyContainer {
-  val videosSearcher = new VideosSearcher
+  private val repository = new InMemoryVideoRepository
+
+  val videosSearcher = new VideosSearcher(repository)
 }
